+++ example/portal/portal.py
@@ -8,7 +8,6 @@ bp_portal.bp_frame.height = 150
 
 bp_portal.render_base = False
 bp_portal.render_hinges = True
-bp_portal.ramp_push = 0
 bp_portal.bp_ramp.width = 10
 bp_portal.make()