+++ example/portal/portal_greebled_ramp_window.py
@@ -15,7 +15,6 @@ bp_portal.bp_ramp.segment_count = 25
 
 bp_portal.render_base = False
 bp_portal.render_hinges = True
-bp_portal.ramp_push = 0
 bp_portal.bp_ramp.width = 10
 bp_portal.make()